sales_flow/pyproject.toml
[project]
name = "sales_flow"
version = "0.1.0"
description = "sales-flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "crewai[tools]>=0.108.0,<1.0.0",
    "mlflow>=2.21.3",
]

[project.scripts]
kickoff = "sales_flow.main:kickoff"
plot = "sales_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
